+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Bugfix
   - fix: `StkTableColumn['width']` not work correctly.
   - fix: when props.showHeaderOverflow not set, the header row height not work correctly in virtual mode .
+  - fix: when `props.border=false`, the border still shown.
 
 ## 0.6.4
 * Feature